Fixed multiple minor bugs and renamed POSIX files
