+++ lib/Core/SpecialFunctionHandler.h
@@ -10,6 +10,8 @@
 #ifndef KLEE_SPECIALFUNCTIONHANDLER_H
 #define KLEE_SPECIALFUNCTIONHANDLER_H
 
+#include "klee/Expr.h"
+
 #include <iterator>
 #include <map>
 #include <vector>
@@ -91,6 +93,8 @@ namespace klee {
     /* Convenience routines */
 
     std::string readStringAtAddress(ExecutionState &state, ref<Expr> address);
+   
+    bool writeConcreteValue(ExecutionState &state, ref<Expr> address, uint64_t value, Expr::Width width);
     
     /* Handlers */
 
@@ -109,9 +113,11 @@ namespace klee {
     HANDLER(handleExit);
     HANDLER(handleAliasFunction);
     HANDLER(handleFree);
+    HANDLER(handleGetContext);
     HANDLER(handleGetErrno);
     HANDLER(handleGetObjSize);
     HANDLER(handleGetValue);
+    HANDLER(handleGetWList);
     HANDLER(handleIsSymbolic);
     HANDLER(handleMakeSymbolic);
     HANDLER(handleMalloc);
