+++ runtime/POSIX/common.h
@@ -33,8 +33,6 @@
 #ifndef COMMON_H_
 #define COMMON_H_
 
-#include "config.h"
-
 #include <sys/types.h>
 #include <string.h>
 #include <stdint.h>
