+++ runtime/POSIX/klee_init_env.c
@@ -12,6 +12,7 @@
 #define _LARGEFILE64_SOURCE
 #endif
 #include "fd.h"
+#include "threads.h"
 
 #include <stdlib.h>
 #include <string.h>
@@ -193,5 +194,6 @@ usage: (klee_init_env) [options] [program arguments]\n\
   klee_init_fds(sym_files, sym_file_len, 
 		sym_stdout_flag, save_all_writes_flag, 
 		fd_fail);
+  klee_init_threads();
 }
