+++ tools/klee-replay/klee-replay.c
@@ -447,6 +447,12 @@ void klee_make_symbolic(void *addr, size_t nbytes, const char *name) {
   }
 }
 
+uint64_t klee_get_wlist(void) {
+      static uint64_t counter = 0;
+
+        return counter++;
+}
+
 /* Redefined here so that we can check the value read. */
 int klee_range(int start, int end, const char* name) {
   int r;
