+++ tools/klee-replay/threads_init.c
@@ -1,0 +1,1 @@
+#include "../../runtime/POSIX/threads_init.c"
