+++ tools/klee/main.cpp
@@ -729,6 +729,7 @@ static const char *modelledExternals[] = {
   "klee_assume", 
   "klee_check_memory_access",
   "klee_define_fixed_object",
+  "klee_get_context",
   "klee_get_errno", 
   "klee_get_valuef",
   "klee_get_valued",
@@ -737,6 +738,7 @@ static const char *modelledExternals[] = {
   "klee_get_value_i32",
   "klee_get_value_i64",
   "klee_get_obj_size", 
+  "klee_get_wlist",
   "klee_is_symbolic", 
   "klee_make_symbolic", 
   "klee_mark_global", 
@@ -751,6 +753,11 @@ static const char *modelledExternals[] = {
   "klee_warning_once", 
   "klee_alias_function",
   "klee_stack_trace",
+  "klee_thread_create",
+  "klee_thread_notify",
+  "klee_thread_preempt",
+  "klee_thread_sleep",
+  "klee_thread_terminate",
 #if LLVM_VERSION_CODE >= LLVM_VERSION(3, 1)
   "llvm.dbg.declare",
   "llvm.dbg.value",
